Updated the create function, changed intro file name
I updated the create function so that it can create a text file to the user's home directory. I also renamed the intro file (just because it was a little long), and also made it into a simple class . That way we can call an instance of it in the actual ProjectMain file. I also made it so that the program will keep running until the user types in "/exit".